move wrapper out of tasks.py
